+++ src/teachings.rs
@@ -1,0 +1,1 @@
+mod teachings;
